use int16_t first for histories
adds saturation. Bench is not changed, but could on int16 overflow.

Bench: 6107863

+++ src/search.cpp
@@ -198,8 +198,8 @@ void Search::clear() {
       th->counterMoveHistory.clear();
       th->resetCalls = true;
       CounterMoveStats& cm = th->counterMoveHistory[NO_PIECE][0];
-      int* t = &cm[NO_PIECE][0];
-      std::fill(t, t + sizeof(cm), CounterMovePruneThreshold - 1);
+      int16_t* t = &cm[NO_PIECE][0];
+      std::fill(t, t + sizeof(cm), int16_t(CounterMovePruneThreshold - 1));
   }
 
   Threads.main()->previousScore = VALUE_INFINITE;
